Rename function to just nasa
- also see if page is rendered now
