+++ eleventy.config.js
@@ -3,7 +3,7 @@ const { EleventyEdgePlugin, EleventyServerlessBundlerPlugin } = require("@11ty/e
 module.exports = function (config) {
   config.addPlugin(EleventyEdgePlugin);
 	config.addPlugin(EleventyServerlessBundlerPlugin, {
-		name: "getNasaImage", // The serverless function name from your permalink object
+		name: "nasa", // The serverless function name from your permalink object
 		functionsDir: "./netlify/functions/",
     redirects: "netlify-toml",
 	});
