+++ netlify.toml
@@ -12,7 +12,7 @@ path = "/"
 
 [[redirects]]
 from = "/"
-to = "/.netlify/functions/getNasaImage"
+to = "/.netlify/functions/nasa"
 status = 200
 force = true
-_generated_by_eleventy_serverless = "getNasaImage"
+_generated_by_eleventy_serverless = "nasa"
