+++ src/index.liquid
@@ -2,7 +2,7 @@
 title: 'Hello, Netlify!'
 layout: 'layouts/base.html'
 permalink:
-  getNasaImage: /
+  nasa: /
 ---
 
 <header>
